pyproject.toml
[project]
name = "kmds-data-helper"
version = "0.1.0"
description = "AI-powered data dictionary generator for KMDS"
requires-python = ">=3.12"
dependencies = [
    "pymupdf4llm>=0.0.12",
    "ollama>=0.1.0",
    "pandas>=2.0.0",
    "setuptools<70.0.0",
    "urllib3<2.0.0",
    "rich>=13.0.0",
    "nbformat>=5.10.4",
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
    "requests>=2.32.5",
    "pytest>=9.0.3",
    "pypdf>=6.11.0",
    "fg-data-profiling>=4.19.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-xdist",
    "requests"
]

[project.scripts]
# Maps kmds-check to your verified, production pre-flight entry point
kmds-check = "kmds_data_helper.kmds_check:main"
# NEW: Maps kmds-analyze to your unified application CLI execution entry point
kmds-analyze = "kmds_data_helper.cli:main"
kmds-test = "pytest:main"

[tool.pytest.ini_options]
# Ensures pytest only runs the new persona tests and stays away from broken legacy files
testpaths = ["tests/test_personas.py"]
python_files = "test_*.py"

# -n 3: Runs 3 parallel workers
# -v: Verbose output
# --disable-warnings: Hides those annoying urllib3 version mismatch logs
addopts = "-n 3 -v --disable-warnings"

filterwarnings = [
    "ignore:urllib3.*match a supported version:requests.RequestsDependencyWarning",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-xdist>=3.8.0",
]
